AI-Voice-Agent/pyproject.toml
[project]
name = "ai-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.7.0",
    "google-genai>=1.57.0",
    "livekit-agents[cartesia,google]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
